-minor bug fix for parserhelper
git-svn-id: https://develop.office.rac.de:4433/svn/OpenSCENARIO/trunk/subprojects/OpenScenarioApi/source/cpp@148 11201c34-f64f-4266-9679-50817c76cc9f

+++ OpenScenarioLib/src/parser/ParserHelper.h
@@ -29,8 +29,8 @@ namespace RAC_OPENSCENARIO
     class ParserHelper 
     {
     public:
-        const static int UNSIGNED_SHORT_MAX_VALUE = USHRT_MAX;
-        const static long UNSIGNED_INT_MAX_VALUE = UINT_MAX;
+        const static long UNSIGNED_SHORT_MAX_VALUE = USHRT_MAX;
+        const static long long UNSIGNED_INT_MAX_VALUE = UINT_MAX;
 
         /**
          * Parsing a string value into a string value.
